refactor: organize styles
