+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import App from "@/App";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ThemeProvider } from "styled-components";
-import { theme } from "@/theme";
+import { theme } from "@/styles";
 
 const qc = new QueryClient();
