--- src/theme.ts
@@ -1,15 +1,0 @@
-export const theme = {
-  colors: {
-    bg: "#f7f6f2",
-    card: "#ffffff",
-    text: "#1f2937",
-    subtle: "#6b7280",
-    border: "#e5e7eb",
-    gold: "#d4af37",
-    goldSoft: "#fff7e0",
-    goldDark: "#b48918",
-  },
-  radius: "14px",
-  shadow: "0 10px 30px rgba(0,0,0,0.08)",
-};
-export type AppTheme = typeof theme;
